さらに針生成処理を見直し
