+++ Assets/Object/Needle/Script/StapedNeedle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19de9af63ca2e054b81c83ab281dc49d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
